Separate Prefork logic
